+++ CHANGELOG.md
@@ -12,10 +12,13 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Initial implementation of shorthand channels
 * Parametarized `max_speed` of toydata objects
 * Add `combine_kwcoco_measures` function
+* Add new API methods to ChannelSpec objects
 
 
 ### Changed
 
+* Reworking kwcoco channel spec backend implementation. Trying to maintain
+  compatibility and insert warnings where things might change.
 * Removed six from CocoDataset
 * Added support for pathlib